Crate diplomat

The Diplomat macro crate.

This crate provides the #[diplomat::bridge] proc macro, which can be used to generate FFI hooks for use with bindings generated by the diplomat-tool library.

Be sure to also include the diplomat_runtime crate when using the #[diplomat::bridge] macro.

Modules

enum_convert 🔒

transparent_convert 🔒

Macros

expose_attrs 🔒

Structs

AttributeInfo 🔒

FuncGen 🔒

Functions

cfgs_to_stream 🔒

gen_bridge 🔒

gen_custom_function 🔒

gen_custom_trait_impl 🔒

gen_custom_type_method 🔒

gen_custom_vtable 🔒

param_conversion 🔒

param_ty 🔒

Attribute Macros

attr

bridge

Mark a module to be exposed through Diplomat-generated FFI.

config

demo

enum_convert

Generate From and Into implementations for a Diplomat enum

macro_rules

opaque

transparent_convert

Generate conversions from inner types for opaque Diplomat types with a single field