Crate diplomat

Modules

enum_convert ğŸ”’

transparent_convert ğŸ”’

Structs

AttributeInfo ğŸ”’

Functions

cfgs_to_stream ğŸ”’

gen_bridge ğŸ”’

gen_custom_type_method ğŸ”’

gen_params_at_boundary ğŸ”’

gen_params_invocation ğŸ”’

Attribute Macros

bridge

Mark a module to be exposed through Diplomat-generated FFI.

enum_convert

Generate From and Into implementations for a Diplomat enum

transparent_convert

Generate conversions from inner types for opaque Diplomat types with a single field